fix(delegation-controller): rename files and fix changelog
